security, authentication

+++ src/main/java/hu/webuni/hr/doravar/web/HolidayController.java
@@ -10,6 +10,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -64,12 +67,15 @@ public class HolidayController {
 	}
 
 	@PostMapping
+	@PreAuthorize("#holidayDto.employeeId == authentication.principal.employee.id") // authentication.principal: HrUser
+																					// employee-ra utal
 	public HolidayDto createHoliday(@RequestBody @Valid HolidayDto holidayDto) {
 		Holiday holiday = holidayService.save(holidayMapper.dtoToHoliday(holidayDto));
 		return holidayMapper.holidayToDto(holiday);
 	}
 
 	@PutMapping("/{id}")
+	@PreAuthorize("#holidayDto.employeeId == authentication.principal.employee.id")
 	public ResponseEntity<HolidayDto> modifyHoliday(@PathVariable long id, @RequestBody @Valid HolidayDto holidayDto) {
 		try {
 			holidayDto.setId(id);
@@ -86,7 +92,12 @@ public class HolidayController {
 	@PutMapping(path = "/approval/{id}", params = "stateOfHolidayRequest")
 	public ResponseEntity<HolidayDto> approveHoliday(@PathVariable long id,
 			@RequestParam String stateOfHolidayRequest) {
+		UserDetails user = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		String username = user.getUsername();
 		try {
+			if (!holidayRepository.findById(id).get().getApprover().getUsername().equals(username)) {
+				throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED);
+			}
 			return ResponseEntity.ok(holidayMapper.holidayToDto(holidayService.approve(id, stateOfHolidayRequest)));
 		} catch (NoSuchElementException e) {
 			return ResponseEntity.notFound().build();
@@ -95,9 +106,14 @@ public class HolidayController {
 		}
 	}
 
-	@DeleteMapping("/{id}")
-	public void deleteHoliday(@PathVariable long id) {
+	@DeleteMapping(path = "/deleteholiday")
+	public void deleteHoliday(@RequestParam(name = "holidayId") long id) {
+		UserDetails user = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		String username = user.getUsername();
 		try {
+			if (!holidayRepository.findById(id).get().getEmployee().getUsername().equals(username)) {
+				throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED);
+			}
 			holidayService.deleteById(id);
 		} catch (NoSuchElementException e) {
 			throw new ResponseStatusException(HttpStatus.NOT_FOUND);
